<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

implement inline projection for sort
1. Planner: prune output columns  in `LogicalSort`
2. Executor: do not copy columns that output doesn't need in `SortExec`

### What is changed and how it works?

```
name                                                   old time/op    new time/op    delta
SortExec/(rows:3000000,_orderBy:[0],_output:[1_2])-12     4.27s ± 3%     4.04s ± 3%   -5.48%  (p=0.016 n=5+5)
SortExec/(rows:3000000,_orderBy:[0],_output:[1])-12       4.26s ± 2%     3.83s ± 3%  -10.02%  (p=0.008 n=5+5)

name                                                   old alloc/op   new alloc/op   delta
SortExec/(rows:3000000,_orderBy:[0],_output:[1_2])-12    27.9MB ± 0%    27.9MB ± 0%   -0.13%  (p=0.008 n=5+5)
SortExec/(rows:3000000,_orderBy:[0],_output:[1])-12      27.9MB ± 0%    27.8MB ± 0%   -0.26%  (p=0.008 n=5+5)

name                                                   old allocs/op  new allocs/op  delta
SortExec/(rows:3000000,_orderBy:[0],_output:[1_2])-12     41.1k ± 0%     41.1k ± 0%   -0.03%  (p=0.029 n=4+4)
SortExec/(rows:3000000,_orderBy:[0],_output:[1])-12       41.1k ± 0%     41.1k ± 0%   -0.07%  (p=0.029 n=4+4)
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
 - Integration test
 - Manual test (add detailed scripts or steps below)
 - No code

Code changes

 - Has exported function/method change
 - Has exported variable/fields change
 - Has interface methods change
 - Has persistent data change

Side effects

 - Possible performance regression
 - Increased code complexity
 - Breaking backward compatibility

Related changes

 - Need to cherry-pick to the release branch
 - Need to update the documentation
 - Need to update the `tidb-ansible` repository

Release note

 - Write release note for bug-fix or new feature.